requirements.txt
boto3
requests
redis
uuid
ffmpeg-python
torch
openai-whisper
pandas
dlib
face_recognition
numpy
scipy
opencv-python
cmake
transformers
